feat: add TAP test driver
TAP via automake requires a TAP driver. this is a very strange piece of code that just floats around
freely under lenient license restrictions. it is now included. it will be purged from the project if
TAP is not requested by the user, but this is not yet implemented

re #2