feat: trigger test support file removal if TAP is not requested
testing with TAP requires its own files and a directory. those files need to be present or absent
based on a `cookiecutter` choice variable, but `cookiecutter` doesn't directly support this kind of
inclusion or exclusion. so the post-run hook now removes all test files if TAP is not requested by
the user. this should mark test support via `make check`; but i'll try out a copy of the project
template to be sure

re #2