fix: add missing autoconf macros
for reasons that invariably involve hidden junk in my build environment, testing this thing out with
a new dev environment led to a `configure` failure due to a missing autoconf archive macro. that is
now resolved. as before, these are under very generous license restrictions and can be included

re #5