docs: update compilation instructions to, you know, work correctly
the previous and straightforward description of the `make/make check/make install` process in the
templated README did not lead to a successful build with the accompanying `conda` build environment.
on investigation, this is because the userspace-defined CPPFLAGS from `configure` had some global
include garbage in it that was interfering with the path resolution for the `boost` headers. this
problem is resolved by invoking `make CPPFLAGS=""` instead of plain `make` for all `make` calls.
this is a formally correct but unpleasant solution that i may revisit in the future

re #4 #11