feat: use post gen hook to add initial set of tracked files
previously the `git_tracking` choice variable caused `git` to be installed in the appropriate
`conda` environment and triggered repo initialization in the post-generation hook. however, the next
step, adding all relevant files, is really annoying to have to do manually. this adds that `git add`
step to the post-gen hook

re #12