build: somewhat improve handling of zlib/bzlib in autotools
so the support for handling bzip2/zlib in autotools is poor, for various reasons. this adds in some
hacked m4 files derived from boost's compiled library support in autoconf archive. the actual
detection and propagation of the LDADD variables (-lz or -lbz2, traditionally) is broken, and should
be patched later; for now, just adding the standard -lbz2 and -lz in Makefile.am when finter is
requested is sufficient for nearly any reasonable application